+++ src/types/wdi5.types.ts
@@ -1,6 +1,7 @@
 import Log from "sap/base/Log"
 import RecordReplay from "sap/ui/test/RecordReplay"
 import { ControlSelector } from "sap/ui/test/RecordReplay"
+import { WDI5Object } from "../lib/wdi5-object"
 
 // // copypasta from
 // // https://stackoverflow.com/questions/41285211/overriding-interface-property-type-defined-in-typescript-d-ts-file/65561287#65561287
@@ -48,6 +49,43 @@ export interface wdi5Config extends WebdriverIO.Config {
          */
         waitForUI5Timeout?: number
     }
+    capabilities: wdi5Capabilites[] | wdi5MultiRemoteCapability
+}
+
+/**
+ * the "wdi5" prefix is to comply with W3C standards
+ */
+export interface wdi5Capabilites extends WebDriver.DesiredCapabilities {
+    "wdi5:authentication"?: BTPAuthenticator | BasicAuthAuthenticator | CustomAuthenticator | Office365Authenticator
+}
+export interface wdi5MultiRemoteCapability {
+    [key: string]: { capabilities: wdi5Capabilites }
+}
+
+export type BTPAuthenticator = {
+    provider: "BTP"
+    usernameSelector?: string
+    passwordSelector?: string
+    submitSelector?: string
+}
+
+export type BasicAuthAuthenticator = {
+    provider: "BasicAuth"
+}
+
+export type CustomAuthenticator = {
+    provider: "custom"
+    usernameSelector: string
+    passwordSelector: string
+    submitSelector: string
+}
+
+export type Office365Authenticator = {
+    provider: "Office365"
+    usernameSelector?: string
+    passwordSelector?: string
+    submitSelector?: string
+    staySignedIn?: boolean
 }
 
 interface wdi5ControlSelector {
@@ -63,6 +101,10 @@ interface wdi5ControlSelector {
      * Name of the control's view parent
      */
     viewName?: string
+    /**
+     * in Fiori Element land, this attribute is used in dynamic UI compositions
+     */
+    viewId?: string
     /**
      * Fully qualified control class name in dot notation, eg: "sap.m.ObjectHeader"
      */
@@ -143,6 +185,16 @@ export interface clientSide_ui5Response {
     className?: string // getControl
     returnType?: string // executeControlMethod
     nonCircularResultObject?: any
+    uuid?: string // uniquie sap.ui.base.Object id
+    object: WDI5Object
+}
+
+export interface clientSide_ui5Object {
+    uuid: string
+    status: wdi5StatusCode
+    aProtoFunctions?: []
+    className?: string
+    object: WDI5Object
 }
 
 /**